Update Screencatcher_Analyst.Modules.AlterTablesAndColumns.txt.vba
Finished the functions that drop columns and insert primary keys.
Sub DropDateFieldsFromDataTables() removes the date columns from the data tables. This was a manual process when copying a database for use on another Hull.

Sub InsertTrialCardNumbers() copies the Trial Card Primary keys from the Final report table into the data tables. This was a manual process that was messy and it didn't need to be.
Added the functions to insert the Primary key values and to Drop the date columns fro the data tables