Add files via upload
This is the third application form. This also shows the tables as a reference.
Added an Events only table set. This is the base line changes of screening and is to be used where there is not sufficient data for more detailed screening change analysis. Fixed some code formatting and removed some commented out code blocks from when I was unit testing individual functions while writing.